Sketch of future ui
